Move npm package out of generator
