+++ src/main/java/tech/jhipster/lite/module/infrastructure/secondary/FileSystemJHipsterModulesRepository.java
@@ -3,10 +3,10 @@ package tech.jhipster.lite.module.infrastructure.secondary;
 import org.springframework.stereotype.Repository;
 import tech.jhipster.lite.common.domain.ProjectFilesReader;
 import tech.jhipster.lite.error.domain.Assert;
-import tech.jhipster.lite.generator.npm.domain.NpmVersions;
 import tech.jhipster.lite.module.domain.JHipsterModuleChanges;
 import tech.jhipster.lite.module.domain.JHipsterModulesRepository;
 import tech.jhipster.lite.module.domain.postaction.JHipsterModuleExecutionContext;
+import tech.jhipster.lite.npm.domain.NpmVersions;
 
 @Repository
 class FileSystemJHipsterModulesRepository implements JHipsterModulesRepository {
