+++ src/main/java/tech/jhipster/lite/npm/package-info.java
@@ -1,0 +1,2 @@
+@tech.jhipster.lite.SharedKernel
+package tech.jhipster.lite.npm;
