+++ src/test/java/tech/jhipster/lite/npm/infrastructure/secondary/FileSystemNpmVersionsTest.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.npm.infrastructure.secondary;
+package tech.jhipster.lite.npm.infrastructure.secondary;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.Mockito.*;
@@ -10,9 +10,9 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import tech.jhipster.lite.UnitTest;
 import tech.jhipster.lite.common.domain.ProjectFilesReader;
-import tech.jhipster.lite.generator.npm.domain.NpmVersion;
-import tech.jhipster.lite.generator.npm.domain.NpmVersionSource;
-import tech.jhipster.lite.generator.npm.domain.UnknownNpmPackageException;
+import tech.jhipster.lite.npm.domain.NpmVersion;
+import tech.jhipster.lite.npm.domain.NpmVersionSource;
+import tech.jhipster.lite.npm.domain.UnknownNpmPackageException;
 
 @UnitTest
 @ExtendWith(MockitoExtension.class)
